# src/pkg/server/handler/apihandler/embedded/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventory Manager",
    "description": "Inventory Manager API",
    "version": "1.0.0",
    "contact": {
      "name": "LinuxCode",
      "email": "jbensch@linuxcode.net"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "inventory.linuxcode.net",
      "description": "Inventory server for example"
    }
  ],
...
  "paths": {
    "/api/items": {
      "get": {
        "summary": "Get all items",
        "description": "Retrieve all items",
        "operationId": "getAllItems",
        "tags": [
          "item"
        ],
        "responses": {
          "200": {
            "description": "successful operation - returns a list of items",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "variant": "rare",
                      "name": "Sword",
                      "text": "A sword that can be used to fight small enemies.",
                      "buy_value": 100,
                      "sell_value": 50,
                      "weight": 10,
                      "durability": 100,
                      "max_stack": 1,
                      "shape": {
                        "rawshape": "###",
                        "height": 3,
                        "width": 1
                      },
                      "type": "meleeWeapon"
                    },
                    {
                      "id": 2,
                      "variant": "common",
                      "name": "Shield",
                      "text": "A shield that can be used to block enemy attacks.",
                      "buy_value": 50,
                      "sell_value": 25,
                      "weight": 15,
                      "durability": 150,
                      "max_stack": 1,
                      "shape": {
                        "rawshape": "####",
                        "height": 2,
                        "width": 2
                      },
                      "type": "meleeWeapon"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "pages": 10,
                    "limit": 10,
                    "total": 100,
                    "offset": 0
                  }
                }
              }
            }
          },
          "404": {
            "description": "No items found"
          }
        },
        "security": []
      },
      "post": {
        "summary": "Add a new item",
        "description": "Add a new item to the system",
        "operationId": "addItem",
        "tags": [
          "item"
        ],
        "requestBody": {
          "required": true,
          "description": "Item to add to the system",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemPostRequest"
              },
              "example": {
                "name": "Sword",
                "variant": "Iron",
                "text": "A sword made of iron.",
                "buy_value": 200,
                "sell_value": 70,
                "weight": 20,
                "durability": 350,
                "max_stack": 1,
                "type": "meleeWeapon",
                "shape": {
                  "rawshape": "###",
                  "height": 3,
                  "width": 1
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added item - returns the newly created item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                },
                "example": {
                  "id": 1,
                  "variant": "gigantic",
                  "name": "Sword",
                  "text": "A sword that can be used to fight gigantic enemies.",
                  "buy_value": 100,
                  "sell_value": 50,
                  "weight": 10,
                  "durability": 100,
                  "max_stack": 1,
                  "shape": {
                    "rawshape": "###",
                    "height": 3,
                    "width": 1
                  },
                  "type": "meleeWeapon"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - e.g. invalid item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/items/{itemId}": {
      "get": {
        "summary": "Get item",
        "description": "Get an item by ID",
        "operationId": "getItemById",
        "tags": [
          "item"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "ID of item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation - returns item",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  },
                  "example": {
                    "id": 1,
                    "variant": "gigantic",
                    "name": "Sword",
                    "text": "A sword that can be used to fight gigantic enemies.",
                    "buy_value": 100,
                    "sell_value": 50,
                    "weight": 10,
                    "durability": 100,
                    "max_stack": 1,
                    "shape": {
                      "rawshape": "###",
                      "height": 3,
                      "width": 1
                    },
                    "type": "meleeWeapon"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID of item supplied (e.g. not an integer) - could not return item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find item (item ID) to return"
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update an item",
        "description": "Update an item by ID",
        "operationId": "updateItemById",
        "tags": [
          "item"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "ID of item that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Item object that needs to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemPostRequest"
              },
              "example": {
                "name": "Sword",
                "variant": "Iron",
                "text": "A sword made of iron.",
                "buy_value": 200,
                "sell_value": 70,
                "weight": 20,
                "durability": 350,
                "max_stack": 1,
                "type": "meleeWeapon",
                "shape": {
                  "rawshape": "###",
                  "height": 3,
                  "width": 1
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update done"
          },
          "400": {
            "description": "Invalid ID of Item supplied (e.g. not an integer) - could not update item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find item (item ID) to update"
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete an item",
        "description": "Delete an item by ID",
        "operationId": "deleteItemById",
        "tags": [
          "item"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "ID of item to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted item (item ID)"
          },
          "400": {
            "description": "Invalid ID of Item supplied (e.g. not an integer) - could not delete item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find item (item ID) to delete"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "ItemShape": {
        "type": "object",
        "description": "This object holds the item shape data.",
        "properties": {
          "rawshape": {
            "type": "string",
            "example": "#####..#####"
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "example": 3
          },
          "width": {
            "type": "integer",
            "format": "int64",
            "example": 4
          }
        },
        "example": {
          "rawshape": "#####..#####",
          "height": 3,
          "width": 4
        },
        "required": [
          "rawshape",
          "height",
          "width"
        ]
      },
      "ItemType": {
        "type": "string",
        "description": "The type of the item.",
        "example": "meleeWeapon",
        "enum": [
          "consumable",
          "armor",
          "rangedWeapon",
          "meleeWeapon",
          "consumableWeapon",
          "quest",
          "resource"
        ]
      },
      "Item": {
        "type": "object",
        "description": "This object holds the item data.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "variant": {
            "type": "string",
            "example": "fast"
          },
          "name": {
            "type": "string",
            "example": "Sword"
          },
          "text": {
            "type": "string",
            "example": "A sword that can be used to fight slow enemies."
          },
          "buy_value": {
            "type": "integer",
            "example": 100
          },
          "sell_value": {
            "type": "integer",
            "example": 50
          },
          "weight": {
            "type": "integer",
            "example": 10
          },
          "durability": {
            "type": "integer",
            "example": 100
          },
          "max_stack": {
            "type": "integer",
            "example": 1
          },
          "shape": {
            "$ref": "#/components/schemas/ItemShape"
          },
          "type": {
            "$ref": "#/components/schemas/ItemType"
          }
        },
        "example": {
          "id": 1,
          "variant": "gigantic",
          "name": "Sword",
          "text": "A sword that can be used to fight gigantic enemies.",
          "buy_value": 100,
          "sell_value": 50,
          "weight": 10,
          "durability": 100,
          "max_stack": 1,
          "shape": {
            "rawshape": "###",
            "height": 3,
            "width": 1
          },
          "type": "meleeWeapon"
        },
        "required": [
          "id",
          "variant",
          "name",
          "text",
          "buy_value",
          "sell_value",
          "weight",
          "durability",
          "max_stack",
          "shape",
          "type"
        ]
      },
      "ItemList": {
        "type": "array",
        "description": "The list of items.",
        "items": {
          "$ref": "#/components/schemas/Item"
        },
        "example": [
          {
            "id": 1,
            "variant": "rare",
            "name": "Sword",
            "text": "A sword that can be used to fight enemies.",
            "buy_value": 100,
            "sell_value": 50,
            "weight": 10,
            "durability": 100,
            "max_stack": 1,
            "shape": {
              "rawshape": "###",
              "height": 3,
              "width": 1
            },
            "type": "meleeWeapon"
          },
          {
            "id": 2,
            "variant": "common",
            "name": "Shield",
            "text": "A shield that can be used to block enemy attacks.",
            "buy_value": 50,
            "sell_value": 25,
            "weight": 15,
            "durability": 150,
            "max_stack": 1,
            "shape": {
              "rawshape": "####",
              "height": 2,
              "width": 2
            },
            "type": "meleeWeapon"
          }
        ]
      },
      "Pagination": {
        "type": "object",
        "description": "This object holds the pagination data.",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "example": 100
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "example": 0
          }
        },
        "example": {
          "page": 1,
          "pages": 10,
          "limit": 10,
          "total": 100,
          "offset": 0
        },
        "required": [
          "page",
          "pages",
          "limit",
          "total",
          "offset"
        ]
      },
      "ItemListResponse": {
        "type": "object",
        "description": "This object holds the item list response data. It includes the list of items and pagination data.",
        "properties": {
          "items": {
            "$ref": "#/components/schemas/ItemList"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "example": {
          "items": [
            {
              "id": 1,
              "variant": "rare",
              "name": "Sword",
              "text": "A sword that can be used to fight small enemies.",
              "buy_value": 100,
              "sell_value": 50,
              "weight": 10,
              "durability": 100,
              "max_stack": 1,
              "shape": {
                "rawshape": "###",
                "height": 3,
                "width": 1
              },
              "type": "meleeWeapon"
            },
            {
              "id": 2,
              "variant": "common",
              "name": "Shield",
              "text": "A shield that can be used to block enemy attacks.",
              "buy_value": 50,
              "sell_value": 25,
              "weight": 15,
              "durability": 150,
              "max_stack": 1,
              "shape": {
                "rawshape": "####",
                "height": 2,
                "width": 2
              },
              "type": "meleeWeapon"
            }
          ],
          "pagination": {
            "page": 1,
            "pages": 10,
            "limit": 10,
            "total": 100,
            "offset": 0
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
      "ItemPostRequest": {
        "type": "object",
        "description": "This object holds the request data for creating a new item.",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sword"
          },
          "variant": {
            "type": "string",
            "example": "Iron"
          },
          "text": {
            "type": "string",
            "example": "A sword made of iron."
          },
          "buy_value": {
            "type": "integer",
            "format": "int64",
            "example": 100
          },
          "sell_value": {
            "type": "integer",
            "format": "int64",
            "example": 50
          },
          "weight": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "durability": {
            "type": "integer",
            "format": "int64",
            "example": 200
          },
          "max_stack": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "type": {
            "$ref": "#/components/schemas/ItemType"
          },
          "shape": {
            "$ref": "#/components/schemas/ItemShape"
          }
        },
        "example": {
          "name": "Sword",
          "variant": "Iron",
          "text": "A sword made of iron.",
          "buy_value": 200,
          "sell_value": 70,
          "weight": 20,
          "durability": 350,
          "max_stack": 1,
          "type": "meleeWeapon",
          "shape": {
            "rawshape": "###",
            "height": 3,
            "width": 1
          }
        },
        "required": [
          "name",
          "variant",
          "text",
          "buy_value",
          "sell_value",
          "weight",
          "durability",
          "max_stack",
          "type",
          "shape"
        ]
      },
      "ErrorType": {
        "type": "string",
        "description": "The error type",
        "enum": [
          "Bad Request",
          "Not Found",
          "Not Implemented",
          "Internal Server Error"
        ],
        "example": "Bad Request"
      },
      "Error": {
        "type": "object",
        "description": "This object holds the error response data.",
        "properties": {
          "errorId": {
            "type": "string",
            "description": "The unique identifier for the error",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "error": {
            "type": "string",
            "description": "The error message",
            "example": "The item does not exist"
          },
          "ErrorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "code": {
            "type": "integer",
            "description": "The error code",
            "example": 404
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The time the error occurred",
            "example": "2020-01-01T00:00:00Z"
          }
        },
        "example": {
          "errorId": "123e4567-e89b-12d3-a456-426614174000",
          "error": "The item does not exist",
          "ErrorType": "Not Found",
          "code": 404,
          "timestamp": "2020-01-01T00:00:00Z"
        },
        "required": [
          "errorId",
          "error",
          "ErrorType",
          "code",
          "timestamp"
        ]
      },
...
    }
  }
}